Moved PydanticObjectType-related code to objecttype package
