+++ graphene_pydantic/__init__.py
@@ -1,3 +1,3 @@
-from .types import PydanticObjectType
+from .objecttype.types import PydanticObjectType
 
 __all__ = ["PydanticObjectType"]
